Fix 32-bit builds: widen the opcode bounds checks
maxOff and minImm/maxImm are sized for the 64-bit opcode layout, so comparing
them against an int failed to compile on 386, arm and wasm32 (TinyGo's wasm
target). Compare in int64; the layout is unchanged.

+++ schema.go
@@ -167,7 +167,7 @@ const (
 )
 
 func makeNode(op Opcode, off, n int) Opcode {
-	if off < 0 || off > maxOff {
+	if off < 0 || int64(off) > maxOff {
 		panic(off)
 	}
 	if n < 0 || n > maxArg {
@@ -186,7 +186,7 @@ func MakeInt(v int64) Opcode {
 }
 
 func makeImm(op Opcode, v int) Opcode {
-	if v < minImm || v > maxImm {
+	if int64(v) < minImm || int64(v) > maxImm {
 		panic(v)
 	}
